feat: enhance LeetCode Interview page and restructure AI Competition
- Introduce AudienceBlock and LearningTimeLine components to the LeetCode Interview page
- Create a new folder hierarchy for AI Competition and LeetCode Interview pages
- Separate data concerns by moving relevant content to data.tsx files for both AI Competition and LeetCode Interview sections
